+++ CHANGELOG.md
@@ -6,6 +6,8 @@ BREAKING CHANGES
 
 - [core] Updated commission handling
 - [core] Fix multisend issue
+- [core] Extend max tx size
+- [api] Fixes in error responses
 
 ## 0.11.0